separating sections by file for website
