+++ code/drasil-website/Website/Drasil_Website.html
@@ -18,7 +18,7 @@
         <h1></h1>
         <div id="Figure:Drasil">
           <figure>
-            <img src="../../images/Icon.png" alt="" width="70.0%" >
+            <img src="../../drasil-website/images/Icon.png" alt="" width="50.0%" >
             <figcaption></figcaption>
           </figure>
         </div>
